# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Purchasing Management API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      responses:
        "200":
          description: Users
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/User" }
    post:
      operationId: createUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateUserInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
  /users/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    patch:
      operationId: updateUser
      tags: [users]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateUserInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/User" }
    delete:
      operationId: deleteUser
      tags: [users]
      responses:
        "204": { description: Deleted }

  # ---------------- DEPARTMENTS ----------------
...
  /workflows:
    get:
      operationId: listWorkflows
      tags: [workflows]
      parameters:
        - { name: departmentId, in: query, required: false, schema: { type: integer } }
        - { name: step, in: query, required: false, schema: { $ref: "#/components/schemas/WorkflowStep" } }
        - { name: q, in: query, required: false, schema: { type: string } }
      responses:
        "200":
          description: Workflows
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/WorkflowSummary" }
    post:
      operationId: createWorkflow
      tags: [workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWorkflowInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
...
  /workflows/{id}:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    get:
      operationId: getWorkflow
      tags: [workflows]
      responses:
        "200":
          description: Workflow
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
    patch:
      operationId: updateWorkflow
      tags: [workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/UpdateWorkflowInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
    delete:
      operationId: deleteWorkflow
      tags: [workflows]
      responses:
        "204": { description: Deleted }
  /workflows/{id}/advance:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    post:
      operationId: advanceWorkflow
      tags: [workflows]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/AdvanceWorkflowInput" }
      responses:
        "200":
          description: Advanced
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }
  /workflows/{id}/undo:
    parameters:
      - { name: id, in: path, required: true, schema: { type: integer } }
    post:
      operationId: undoWorkflow
      tags: [workflows]
      responses:
        "200":
          description: Reverted
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Workflow" }

  # ---------------- DOCUMENTS ----------------
...
components:
  schemas:
...
    Role:
      type: string
      enum:
        - ADMIN
        - FINANCIAL_ALL
        - FINANCIAL_INVOICE
        - FINANCIAL_PAYMENT
        - DEPT_MANAGER
        - DEPT_USER
        - GT_INVEST
        - READ_ONLY_DEPT
        - READ_ONLY_ALL
    WorkflowStep:
      type: string
      enum:
        - NEW
        - QUOTATION
        - VALIDATING_QUOTE_FINANCIAL
        - VALIDATING_BY_FINANCIAL
        - GT_INVEST
        - ORDERING
        - DELIVERY
        - INVOICE
        - VALIDATING_INVOICE
        - PAYMENT
        - DONE
    Priority:
      type: string
      enum: [LOW, NORMAL, HIGH, URGENT]
...
    User:
      type: object
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email: { type: ["string", "null"] }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        departmentIds:
          type: array
          items: { type: integer }
        source: { type: string, enum: [LOCAL, LDAP] }
        createdAt: { type: string, format: date-time }
      required: [id, username, displayName, roles, departmentIds, source, createdAt]
    CreateUserInput:
      type: object
      properties:
        username: { type: string }
        displayName: { type: string }
        email: { type: ["string", "null"] }
        password: { type: ["string", "null"] }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        departmentIds:
          type: array
          items: { type: integer }
      required: [username, displayName, roles, departmentIds]
    UpdateUserInput:
      type: object
      properties:
        displayName: { type: ["string", "null"] }
        email: { type: ["string", "null"] }
        password: { type: ["string", "null"] }
        roles:
          type: array
          items: { $ref: "#/components/schemas/Role" }
        departmentIds:
          type: array
          items: { type: integer }
...
    QuoteEntry:
      type: object
      properties:
        companyId: { type: ["integer", "null"] }
        companyName: { type: ["string", "null"] }
        contactId: { type: ["integer", "null"] }
        amount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        notes: { type: ["string", "null"] }
        winning: { type: boolean }
      required: [winning]
    Workflow:
      type: object
      properties:
        id: { type: integer }
        reference: { type: string }
        title: { type: string }
        departmentId: { type: integer }
        departmentName: { type: string }
        createdById: { type: integer }
        createdByName: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        currentStep: { $ref: "#/components/schemas/WorkflowStep" }
        branch: { type: ["string", "null"], enum: [K_ORDER, GT_INVEST, null] }
        ageDays: { type: integer }
        isStalled: { type: boolean }
        # Step 1 - New
        description: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        estimatedAmount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        neededBy: { type: ["string", "null"], format: date }
        # Step 2 - Quotation
        quotes:
          type: array
          items: { $ref: "#/components/schemas/QuoteEntry" }
        threeQuoteRequired: { type: boolean }
        # Step 3 - Manager validation
        managerApproved: { type: ["boolean", "null"] }
        managerComment: { type: ["string", "null"] }
        # Step 4 - Financial validation
        financialApproved: { type: ["boolean", "null"] }
        financialComment: { type: ["string", "null"] }
        # GT Invest
        gtInvestDateId: { type: ["integer", "null"] }
        gtInvestResultId: { type: ["integer", "null"] }
        gtInvestComment: { type: ["string", "null"] }
        # Step 5 - Ordering
        orderNumber: { type: ["string", "null"] }
        orderDate: { type: ["string", "null"], format: date }
        # Step 6 - Delivery
        deliveredOn: { type: ["string", "null"], format: date }
        deliveryNotes: { type: ["string", "null"] }
        # Step 7 - Invoice
        invoiceNumber: { type: ["string", "null"] }
        invoiceAmount: { type: ["number", "null"] }
        invoiceDate: { type: ["string", "null"], format: date }
        # Step 8 - Validating Invoice
        invoiceValidated: { type: ["boolean", "null"] }
        invoiceSignedBy: { type: ["string", "null"] }
        invoiceSignedAt: { type: ["string", "null"], format: date-time }
        # Step 9 - Payment
        paymentDate: { type: ["string", "null"], format: date }
        paymentReference: { type: ["string", "null"] }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - reference
        - title
        - departmentId
        - departmentName
        - createdById
        - createdByName
        - priority
        - currentStep
        - ageDays
        - isStalled
        - quotes
        - threeQuoteRequired
        - createdAt
        - updatedAt
    WorkflowSummary:
      type: object
      properties:
        id: { type: integer }
        reference: { type: string }
        title: { type: string }
        departmentId: { type: integer }
        departmentName: { type: string }
        priority: { $ref: "#/components/schemas/Priority" }
        currentStep: { $ref: "#/components/schemas/WorkflowStep" }
        branch: { type: ["string", "null"], enum: [K_ORDER, GT_INVEST, null] }
        estimatedAmount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        ageDays: { type: integer }
        isStalled: { type: boolean }
        createdByName: { type: string }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
      required:
        - id
        - reference
        - title
        - departmentId
        - departmentName
        - priority
        - currentStep
        - ageDays
        - isStalled
        - createdByName
        - createdAt
        - updatedAt
...
    CreateWorkflowInput:
      type: object
      properties:
        title: { type: string }
        departmentId: { type: integer }
        priority: { $ref: "#/components/schemas/Priority" }
        description: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        estimatedAmount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        neededBy: { type: ["string", "null"], format: date }
      required: [title, departmentId, priority]
    UpdateWorkflowInput:
      type: object
      properties:
        title: { type: ["string", "null"] }
        priority: { $ref: "#/components/schemas/Priority" }
        description: { type: ["string", "null"] }
        category: { type: ["string", "null"] }
        estimatedAmount: { type: ["number", "null"] }
        currency: { type: ["string", "null"] }
        neededBy: { type: ["string", "null"], format: date }
        quotes:
          type: array
          items: { $ref: "#/components/schemas/QuoteEntry" }
        managerApproved: { type: ["boolean", "null"] }
        managerComment: { type: ["string", "null"] }
        financialApproved: { type: ["boolean", "null"] }
        financialComment: { type: ["string", "null"] }
        gtInvestDateId: { type: ["integer", "null"] }
        gtInvestResultId: { type: ["integer", "null"] }
        gtInvestComment: { type: ["string", "null"] }
        orderNumber: { type: ["string", "null"] }
        orderDate: { type: ["string", "null"], format: date }
        deliveredOn: { type: ["string", "null"], format: date }
        deliveryNotes: { type: ["string", "null"] }
        invoiceNumber: { type: ["string", "null"] }
        invoiceAmount: { type: ["number", "null"] }
        invoiceDate: { type: ["string", "null"], format: date }
        invoiceValidated: { type: ["boolean", "null"] }
        invoiceSignedBy: { type: ["string", "null"] }
        paymentDate: { type: ["string", "null"], format: date }
        paymentReference: { type: ["string", "null"] }
    AdvanceWorkflowInput:
      type: object
      properties:
        branch: { type: ["string", "null"], enum: [K_ORDER, GT_INVEST, null] }
...